+++ Source/IGListDiffKit/IGListExperiments.h
@@ -20,8 +20,6 @@ typedef NS_OPTIONS (NSInteger, IGListExperiment) {
     IGListExperimentBackgroundDiffing = 1 << 2,
     /// Test fallback to reloadData when "too many" update operations.
     IGListExperimentReloadDataFallback = 1 << 3,
-    /// Test removing the layout pass when calling scrollToObject to avoid creating off-screen cells.
-    IGListExperimentAvoidLayoutOnScrollToObject = 1 << 4,
     /// Test fixing a crash when inserting and deleting the same NSIndexPath multiple times.
     IGListExperimentFixIndexPathImbalance = 1 << 5,
     /// Test deferring object creation until just before diffing.